requirements.txt
# =============================================================================
# QUANTUM LIBRARY RAG - Dependencias
# =============================================================================

# -----------------------------------------------------------------------------
# Framework RAG
# -----------------------------------------------------------------------------
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-vector-stores-qdrant>=0.2.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-llms-anthropic>=0.1.0

# -----------------------------------------------------------------------------
# Vector Database
# -----------------------------------------------------------------------------
qdrant-client>=1.7.0

# -----------------------------------------------------------------------------
# Búsqueda Léxica
# -----------------------------------------------------------------------------
rank-bm25>=0.2.2

# -----------------------------------------------------------------------------
# Grafos
# -----------------------------------------------------------------------------
networkx>=3.2

# -----------------------------------------------------------------------------
# LLM Providers
# -----------------------------------------------------------------------------
openai>=1.10.0
anthropic>=0.18.0

# -----------------------------------------------------------------------------
# Claude Max Client (suscripcion Max, sin coste por token)
# -----------------------------------------------------------------------------
claude-agent-sdk>=0.1.0
anyio>=4.0.0

# -----------------------------------------------------------------------------
# Embeddings Locales (opcional)
# -----------------------------------------------------------------------------
sentence-transformers>=2.2.0
# torch>=2.0.0  # Descomentar si usas embeddings locales

# -----------------------------------------------------------------------------
# Procesamiento de Texto
# -----------------------------------------------------------------------------
tiktoken>=0.5.0  # Tokenización para contar tokens
markdown>=3.5
beautifulsoup4>=4.12.0
regex>=2023.0

# -----------------------------------------------------------------------------
# Utilidades
# -----------------------------------------------------------------------------
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0  # Para output bonito en CLI
click>=8.1.0  # CLI framework
tqdm>=4.66.0  # Progress bars

# -----------------------------------------------------------------------------
# Ejecución de Código (Sandbox)
# -----------------------------------------------------------------------------
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.8.0
sympy>=1.12
pandas>=2.1.0
# qutip>=4.7.0  # Descomentar si necesitas simulaciones cuánticas

# -----------------------------------------------------------------------------
# Logging y Monitorización
# -----------------------------------------------------------------------------
structlog>=23.0.0

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0

# -----------------------------------------------------------------------------
# Desarrollo
# -----------------------------------------------------------------------------
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
